MCP for nonrelation db/requirements.txt
fastapi
uvicorn
pymongo
redis
cassandra-driver
elasticsearch
pandas
numpy
matplotlib
seaborn
plotly
python-dotenv
pydantic
openai
anthropic
langchain
langchain-openai
langchain-anthropic
jinja2
aiofiles
python-multipart